+++ store/data/store.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia"
 
-export const useEventsStore = defineStore(
-  "events",
+export const useDataStore = defineStore(
+  "data",
   () => {
     const version = $ref(0)
