+++ store/data/types/class.ts
