+++ utils/googleSheetParser/helpers/__tests__/mocks/classes.ts
@@ -1,4 +1,4 @@
-import { ClassRawData } from "~/store/events"
+import { ClassRawData } from "~/store/data"
 
 export const MOCK_CLASSES_DATA = `/*O_o*/
 google.visualization.Query.setResponse({"version":"0.6","reqId":"0","status":"ok","sig":"2130838110","table":{"cols":[{"id":"A","label":"","type":"string"},{"id":"B","label":"","type":"string"},{"id":"C","label":"","type":"string"},{"id":"D","label":"","type":"string"},{"id":"E","label":"","type":"string"},{"id":"F","label":"","type":"string"},{"id":"G","label":"","type":"string"},{"id":"H","label":"","type":"string"},{"id":"I","label":"","type":"string"},{"id":"J","label":"","type":"string"},{"id":"K","label":"","type":"string"},{"id":"L","label":"","type":"string"},{"id":"M","label":"","type":"string"},{"id":"N","label":"","type":"string"},{"id":"O","label":"","type":"string"},{"id":"P","label":"","type":"string"},{"id":"Q","label":"","type":"string"},{"id":"R","label":"","type":"string"},{"id":"S","label":"","type":"string"},{"id":"T","label":"","type":"string"},{"id":"U","label":"","type":"string"},{"id":"V","label":"","type":"string"},{"id":"W","label":"","type":"string"},{"id":"X","label":"","type":"string"},{"id":"Y","label":"","type":"string"}],"rows":[{"c":[{"v":"id"},{"v":"subjectId"},{"v":"groups"},{"v":"time"},{"v":"room"},{"v":"note"},{"v":"dates"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"1"},{"v":"protIP"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"215"},{"v":"-"},{"v":"21,28.02.2022 / 14,21,28.03.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"2"},{"v":"retail"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"512"},{"v":"-"},{"v":"04,11,18,25.04.2022 / 16,23,30.05.2022 / 06.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"3"},{"v":"neuro2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"215"},{"v":"-"},{"v":"15,22.02.2022 / 01,15,22,29.03.2022 / 05.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"4"},{"v":"brand"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"215"},{"v":"-"},{"v":"12,19,26.04.2022 / 10,17,24,31.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"5"},{"v":"lead1"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"215"},{"v":"-"},{"v":"05.05.2022 / 03,07.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"6"},{"v":"eng"},{"v":"1"},{"v":"19:00 - 20:20"},{"v":"517"},{"v":"-"},{"v":"09,16.02.2022 / 02,09,16,23,30.03.2022 / 06,13,20,27.04.2022 / 04,11,18,25.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"7"},{"v":"eng"},{"v":"2"},{"v":"19:00 - 20:20"},{"v":"518"},{"v":"-"},{"v":"09,16.02.2022 / 02,09,16,23,30.03.2022 / 06,13,20,27.04.2022 / 04,11,18,25.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"8"},{"v":"eng"},{"v":"3"},{"v":"20:30 - 21:50"},{"v":"517"},{"v":"-"},{"v":"09,16.02.2022 / 02,09,16,23,30.03.2022 / 06,13,20,27.04.2022 / 04,11,18,25.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"9"},{"v":"eng"},{"v":"4"},{"v":"20:30 - 21:50"},{"v":"518"},{"v":"-"},{"v":"09,16.02.2022 / 02,09,16,23,30.03.2022 / 06,13,20,27.04.2022 / 04,11,18,25.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"10"},{"v":"internet"},{"v":"1,2"},{"v":"20:30 - 21:50"},{"v":"519"},{"v":"-"},{"v":"16.02.2022 / 02,09,16,23,30.03.2022 / 06,13,20,27.04.2022 / 04,11,18,25.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"11"},{"v":"internet"},{"v":"3,4"},{"v":"19:00 - 20:20"},{"v":"519"},{"v":"-"},{"v":"16.02.2022 / 02,09,16,23,30.03.2022 / 06,13,20,27.04.2022 / 04,11,18,25.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"12"},{"v":"finance2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"214"},{"v":"-"},{"v":"01,08.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"13"},{"v":"enterp"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"115"},{"v":"-"},{"v":"17,25.02.2022 / 03,04,11,17,31.03.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"14"},{"v":"lux"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"206"},{"v":"-"},{"v":"17.02.2022 / 03,17,31.03.2022 / 14,28.04.2022 / 19.05.2022 / 02.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"15"},{"v":"digital"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"215"},{"v":"-"},{"v":"10,24.02.2022 / 10,24.03.2022 / 07,21.04.2022 / 12,26.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"16"},{"v":"hospit"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"206"},{"v":"-"},{"v":"11,18,25.02.2022 / 04,11,18,25.03.2022 / 01.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"17"},{"v":"iNeuro"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"206"},{"v":"-"},{"v":"08,22,29.04.2022 / 06,20.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"18"},{"v":"entertain"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"207"},{"v":"-"},{"v":"08,15,22,29.04.2022 / 06,13,20,27.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"19"},{"v":"eng"},{"v":"1"},{"v":"09:00 - 11:55"},{"v":"209"},{"v":"-"},{"v":"12,26.02.2022 / 12,26.03.2022 / 09.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"20"},{"v":"eng"},{"v":"2"},{"v":"09:00 - 11:55"},{"v":"210"},{"v":"-"},{"v":"12,26.02.2022 / 12,26.03.2022 / 09.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"21"},{"v":"eng"},{"v":"3"},{"v":"09:00 - 11:55"},{"v":"209"},{"v":"-"},{"v":"19.02.2022 / 05,19.03.2022 / 16.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"22"},{"v":"eng"},{"v":"4"},{"v":"09:00 - 11:55"},{"v":"210"},{"v":"-"},{"v":"19.02.2022 / 05,19.03.2022 / 16.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"23"},{"v":"eng"},{"v":"3"},{"v":"09:00 - 10:15"},{"v":"209"},{"v":"-"},{"v":"02.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"24"},{"v":"eng"},{"v":"4"},{"v":"09:00 - 10:15"},{"v":"210"},{"v":"-"},{"v":"02.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"25"},{"v":"enterp"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"115"},{"v":"По окончании - индивидуальный график"},{"v":"12,26.02.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"26"},{"v":"enterp"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"407"},{"v":"-"},{"v":"19.02.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"27"},{"v":"iNeuro"},{"v":"-"},{"v":"12:55 - 16:00"},{"v":"407"},{"v":"-"},{"v":"09,30.04.2022 / 21.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"28"},{"v":"lead2"},{"v":"-"},{"v":"3 пары"},{"v":"215"},{"v":"-"},{"v":"19.03.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"29"},{"v":"lead3"},{"v":"-"},{"v":"в 15:00"},{"v":"215"},{"v":"онлайн, 3 пары"},{"v":"18.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"30"},{"v":"finance2"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"215"},{"v":"-"},{"v":"05,12.03.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"31"},{"v":"protIP"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"215"},{"v":"-"},{"v":"26.03.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"32"},{"v":"sales"},{"v":"-"},{"v":"10:35 - 17:20"},{"v":"215"},{"v":"лекция"},{"v":"02.04.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"33"},{"v":"sales"},{"v":"1,2"},{"v":"10:35 - 17:20"},{"v":"115"},{"v":"семинар"},{"v":"16.04.2022 / 14.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"34"},{"v":"sales"},{"v":"3,4"},{"v":"10:35 - 17:20"},{"v":"115"},{"v":"семинар"},{"v":"23.04.2022 / 28.05.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"35"},{"v":"exEng"},{"v":"-"},{"v":"с 10 утра"},{"v":"-"},{"v":"-"},{"v":"09,10.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"36"},{"v":"exNeuro"},{"v":"-"},{"v":"с 10 утра"},{"v":"-"},{"v":"-"},{"v":"15.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"37"},{"v":"exIntern"},{"v":"-"},{"v":"с 10 утра"},{"v":"-"},{"v":"-"},{"v":"28.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"38"},{"v":"exIP"},{"v":"-"},{"v":"с 12 утра"},{"v":"-"},{"v":"-"},{"v":"25.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"39"},{"v":"exFin"},{"v":"-"},{"v":"с 10 утра"},{"v":"-"},{"v":"-"},{"v":"20.06.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"40"},{"v":"proj"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"5184"},{"v":"лекция"},{"v":"05,12,19,26.09.2022 / 03,10.10.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"41"},{"v":"proj"},{"v":"1,2"},{"v":"19:00 - 21:50"},{"v":"5184"},{"v":"семинар"},{"v":"17,31.10.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"42"},{"v":"proj"},{"v":"3,4"},{"v":"19:00 - 21:50"},{"v":"5184"},{"v":"семинар"},{"v":"24.10.2022 / 07.11.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"43"},{"v":"web"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"519"},{"v":"-"},{"v":"14,21,28.11.2022 / 05,12.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"44"},{"v":"eng"},{"v":"1"},{"v":"19:00 - 21:50"},{"v":"509"},{"v":"-"},{"v":"06,20.09.2022 / 04,18.10.2022 / 01,15,29.11.2022 / 13.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"45"},{"v":"eng"},{"v":"3"},{"v":"19:00 - 21:50"},{"v":"509"},{"v":"-"},{"v":"13,27.09.2022 / 11,25.10.2022 / 08,22.11.2022 / 06,20.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"46"},{"v":"iNeuro2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"112"},{"v":"-"},{"v":"07,21.09.2022 / 05,19.10.2022 / 02,16,30.11.2022 / 14.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"47"},{"v":"entertain2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"130"},{"v":"-"},{"v":"07,21.09.2022 / 05,19.10.2022 / 02,16,30.11.2022 / 14.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"48"},{"v":"enterp2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"115"},{"v":"-"},{"v":"08,22.09.2022 / 06,20.10.2022 / 03,17.11.2022 / 01,15.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"49"},{"v":"retail2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"130"},{"v":"-"},{"v":"15,29.09.2022 / 13,27.10.2022 / 10,24.11.2022 / 08,22.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"50"},{"v":"digital2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"207"},{"v":"-"},{"v":"09,23.09.2022 / 07,21.10.2022 / 11,18,25.11.2022 / 09,23.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"51"},{"v":"hospit2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"207"},{"v":"-"},{"v":"16,30.09.2022 / 14,28.10.2022 / 18.11.2022 / 02,16.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"52"},{"v":"eng"},{"v":"1"},{"v":"09:00 - 11:55"},{"v":"4054 и 1068"},{"v":"1-я пара в 4054, 2-я пара в 1068"},{"v":"03,17.09.2022 / 01,15,29.10.2022 / 12,26.11.2022 / 10,24.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"53"},{"v":"eng"},{"v":"3"},{"v":"09:00 - 11:55"},{"v":"4054 и 1068"},{"v":"1-я пара в 4054, 2-я пара в 1068"},{"v":"10,24.09.2022 / 08,22.10.2022 / 05,19.11.2022 / 03,17.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"54"},{"v":"adv"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"3030"},{"v":"-"},{"v":"03,10,17,24.09.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"55"},{"v":"qual"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"3030"},{"v":"Внимание: в каждый из дней аудитрия разная"},{"v":"01.10.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"56"},{"v":"eng"},{"v":"2"},{"v":"19:00 - 21:50"},{"v":"510"},{"v":"-"},{"v":"06,20.09.2022 / 04,18.10.2022 / 01,15,29.11.2022 / 13.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"57"},{"v":"eng"},{"v":"4"},{"v":"19:00 - 21:50"},{"v":"510"},{"v":"-"},{"v":"13,27.09.2022 / 11,25.10.2022 / 08,22.11.2022 / 06,20.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"58"},{"v":"lux2"},{"v":"-"},{"v":"19:00 - 21:50"},{"v":"130"},{"v":"-"},{"v":"14,28.09.2022 / 12,26.10.2022 / 09,23.11.2022 / 07,21.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"59"},{"v":"eng"},{"v":"2"},{"v":"09:00 - 11:55"},{"v":"4051 и 3056"},{"v":"1-я пара в 4051, 2-я пара в 3056"},{"v":"03,17.09.2022 / 01,15,29.10.2022 / 12,26.11.2022 / 10,24.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"60"},{"v":"eng"},{"v":"4"},{"v":"09:00 - 11:55"},{"v":"4051 и 3056"},{"v":"1-я пара в 4051, 2-я пара в 3056"},{"v":"10,24.09.2022 / 08,22.10.2022 / 05,19.11.2022 / 03,17.12.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"61"},{"v":"qual"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"3019"},{"v":"Внимание: в каждый из дней аудитрия разная"},{"v":"08.10.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"62"},{"v":"qual"},{"v":"-"},{"v":"12:55 - 17:20"},{"v":"314"},{"v":"Внимание: в каждый из дней аудитрия разная"},{"v":"15.10.2022"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]}],"parsedNumHeaders":0}});`
