+++ utils/googleSheetParser/helpers/__tests__/mocks/specialties.ts
@@ -1,4 +1,4 @@
-import { SpecialtyRawData } from "~/store/events"
+import { SpecialtyRawData } from "~/store/data"
 
 export const MOCK_SPECIALTIES_DATA = `/*O_o*/
 google.visualization.Query.setResponse({"version":"0.6","reqId":"0","status":"ok","sig":"1565345499","table":{"cols":[{"id":"A","label":"","type":"string"},{"id":"B","label":"","type":"string"},{"id":"C","label":"","type":"string"},{"id":"D","label":"","type":"string"},{"id":"E","label":"","type":"string"},{"id":"F","label":"","type":"string"},{"id":"G","label":"","type":"string"},{"id":"H","label":"","type":"string"},{"id":"I","label":"","type":"string"},{"id":"J","label":"","type":"string"},{"id":"K","label":"","type":"string"},{"id":"L","label":"","type":"string"},{"id":"M","label":"","type":"string"},{"id":"N","label":"","type":"string"},{"id":"O","label":"","type":"string"},{"id":"P","label":"","type":"string"},{"id":"Q","label":"","type":"string"},{"id":"R","label":"","type":"string"},{"id":"S","label":"","type":"string"},{"id":"T","label":"","type":"string"},{"id":"U","label":"","type":"string"},{"id":"V","label":"","type":"string"},{"id":"W","label":"","type":"string"},{"id":"X","label":"","type":"string"},{"id":"Y","label":"","type":"string"},{"id":"Z","label":"","type":"string"}],"rows":[{"c":[{"v":"id"},{"v":"title"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"DIGITAL"},{"v":"Digital-маркетинг"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"NEURO"},{"v":"Нейромаркетинг"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"ENTERTAIN"},{"v":"Маркетинг в индустрии развлечений"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"LUX"},{"v":"Маркетинг товаров категории люкс"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"RETAIL"},{"v":"Маркетинг розничных сетей и FMCG"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"HOSPITALITY"},{"v":"Маркетинг услуг в индустрии гостеприимства"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"ENTERPRENEUR"},{"v":"Предпринимательство"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"SPORT"},{"v":"Маркетинг в индустрии спорта"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]}],"parsedNumHeaders":0}});`
