+++ utils/googleSheetParser/helpers/__tests__/mocks/subjects.ts
@@ -1,4 +1,4 @@
-import { SubjectRawData } from "~/store/events"
+import { SubjectRawData } from "~/store/data"
 
 export const MOCK_SUBJECTS_DATA = `/*O_o*/
 google.visualization.Query.setResponse({"version":"0.6","reqId":"0","status":"ok","sig":"1889421812","table":{"cols":[{"id":"A","label":"","type":"string"},{"id":"B","label":"","type":"string"},{"id":"C","label":"","type":"string"},{"id":"D","label":"","type":"string"},{"id":"E","label":"","type":"string"},{"id":"F","label":"","type":"string"},{"id":"G","label":"","type":"string"},{"id":"H","label":"","type":"string"},{"id":"I","label":"","type":"string"},{"id":"J","label":"","type":"string"},{"id":"K","label":"","type":"string"},{"id":"L","label":"","type":"string"},{"id":"M","label":"","type":"string"},{"id":"N","label":"","type":"string"},{"id":"O","label":"","type":"string"},{"id":"P","label":"","type":"string"},{"id":"Q","label":"","type":"string"},{"id":"R","label":"","type":"string"},{"id":"S","label":"","type":"string"},{"id":"T","label":"","type":"string"},{"id":"U","label":"","type":"string"},{"id":"V","label":"","type":"string"},{"id":"W","label":"","type":"string"}],"rows":[{"c":[{"v":"id"},{"v":"title"},{"v":"professor"},{"v":"specs"},{"v":"color"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"finance2"},{"v":"Финансовое моделирование II"},{"v":"Шаренков Сергей Борисович"},{"v":"-"},{"v":"#3399FF"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"protIP"},{"v":"Практика защиты интеллектуальной собственности"},{"v":"Абдуриамова Дженнет Агамурадовна"},{"v":"-"},{"v":"#99CC66"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"retail"},{"v":"Введение в специализацию \u0027Маркетинг розничных сетей\u0027"},{"v":"Филатов Алексей Аркадьевич"},{"v":"RETAIL"},{"v":"#6699CC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"neuro2"},{"v":"Нейромаркетинг II"},{"v":"Клепиков Олег Евгеньевич"},{"v":"-"},{"v":"#9966FF"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"brand"},{"v":"Создание и оценка бренда"},{"v":"Муразанов Алексей Викторович"},{"v":"-"},{"v":"#CC6699"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"eng"},{"v":"Английский"},{"v":"-"},{"v":"-"},{"v":"#FF9966"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"internet"},{"v":"Интернет-маркетинг"},{"v":"Слесарев Максим Александрович"},{"v":"-"},{"v":"#00CCCC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"enterp"},{"v":"Введение в специализацию \u0027Предпринимательство\u0027"},{"v":"Хотяшева Ольга Михайловна"},{"v":"ENTERPRENEUR"},{"v":"#6699CC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"lux"},{"v":"Введение в специализацию \u0027Маркетинг товаров категории люкс\u0027"},{"v":"Дадашева Эльнара Саладдиновна"},{"v":"LUX"},{"v":"#6699CC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"digital"},{"v":"Введение в специализацию \u0027Digital-маркетинг\u0027"},{"v":"Слесарев Максим Александрович"},{"v":"DIGITAL"},{"v":"#6699CC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"hospit"},{"v":"Введение в специализацию \u0027Маркетинг в индустрии гостепреимства\u0027"},{"v":"Кисеев Иван Дмитриевич"},{"v":"HOSPITALITY"},{"v":"#6699CC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"iNeuro"},{"v":"Введение в специализацию \u0027Отраслевой нейромаркетинг\u0027"},{"v":"Клепиков Олег Евгеньевич"},{"v":"NEURO"},{"v":"#6699CC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"entertain"},{"v":"Введение в специализацию \u0027Маркетинг в индустрии развлечений\u0027"},{"v":"Бельская Анастасия Александровна"},{"v":"ENTERTAIN"},{"v":"#6699CC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"sales"},{"v":"Управление продажами"},{"v":"Хотяшева Ольга Михайловна"},{"v":"-"},{"v":"#FF6666"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"lead1"},{"v":"Leadership - Hays"},{"v":"-"},{"v":"-"},{"v":"-"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"lead2"},{"v":"Leadership"},{"v":"Tatiana Rowson"},{"v":"-"},{"v":"-"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"lead3"},{"v":"Leadership"},{"v":"Yulia McVeigh"},{"v":"-"},{"v":"-"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"exEng"},{"v":"Экзамен - Английский"},{"v":"-"},{"v":"-"},{"v":"#32CD32"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"exNeuro"},{"v":"Экзамен - Нейромаркетинг"},{"v":"-"},{"v":"-"},{"v":"#32CD32"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"exIntern"},{"v":"Экзамен - Интернет-маркетинг"},{"v":"-"},{"v":"-"},{"v":"#32CD32"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"exIP"},{"v":"Экзамен - Интеллектуальная собственность"},{"v":"-"},{"v":"-"},{"v":"#32CD32"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"exFin"},{"v":"Экзамен - Финансовое моделирование"},{"v":"-"},{"v":"-"},{"v":"#32CD32"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"proj"},{"v":"Управление проектами"},{"v":"Аншина Марина Львовна"},{"v":"-"},{"v":"#FF6666"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"web"},{"v":"Веб-аналитика"},{"v":"Слесарев Максим Александрович"},{"v":"-"},{"v":"#00CCCC"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"adv"},{"v":"Современные технологии в рекламе"},{"v":"Щелина Лидия Алексеевна"},{"v":"-"},{"v":"#CC6699"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"qual"},{"v":"Управление качеством"},{"v":"Барменков Евгений Юрьевич"},{"v":"-"},{"v":"#99CC66"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"retail2"},{"v":"Специализация \u0027Маркетинг розничных сетей\u0027"},{"v":"Филатов Алексей Аркадьевич"},{"v":"RETAIL"},{"v":"#9B59B6"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"enterp2"},{"v":"Специализация \u0027Предпринимательство\u0027"},{"v":"Хотяшева Ольга Михайловна"},{"v":"ENTERPRENEUR"},{"v":"#9B59B6"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"lux2"},{"v":"Специализация \u0027Маркетинг товаров категории люкс\u0027"},{"v":"Дадашева Эльнара Саладдиновна"},{"v":"LUX"},{"v":"#9B59B6"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"digital2"},{"v":"Специализация \u0027Digital-маркетинг\u0027"},{"v":"Слесарев Максим Александрович"},{"v":"DIGITAL"},{"v":"#9B59B6"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"hospit2"},{"v":"Специализация \u0027Маркетинг в индустрии гостепреимства\u0027"},{"v":"Кисеев Иван Дмитриевич"},{"v":"HOSPITALITY"},{"v":"#9B59B6"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"iNeuro2"},{"v":"Специализация \u0027Отраслевой нейромаркетинг\u0027"},{"v":"Клепиков Олег Евгеньевич"},{"v":"NEURO"},{"v":"#9B59B6"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]},{"c":[{"v":"entertain2"},{"v":"Специализация \u0027Маркетинг в индустрии развлечений\u0027"},{"v":"Бельская Анастасия Александровна"},{"v":"ENTERTAIN"},{"v":"#9B59B6"},null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,{"v":null}]}],"parsedNumHeaders":0}});`
